requirements.txt
# 데이터 분석 및 처리
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# 시각화
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# 데이터베이스 및 파일 처리
sqlite3
json5>=0.9.0

# 설정 파일 처리
pyyaml>=6.0
configparser

# 날짜/시간 처리
python-dateutil>=2.8.0

# 테스트
pytest>=7.0.0
pytest-cov>=4.0.0

# 로깅
colorlog>=6.0.0

# 웹 대시보드 (선택사항)
dash>=2.0.0
dash-bootstrap-components>=1.0.0

# 데이터 검증
pydantic>=1.9.0

# 비동기 처리
asyncio
aiofiles>=0.8.0
